move app insights components to availability
