+++ app/availability/app-insights/app-insights-settings.component.html
