+++ app/availability/app-insights/app-insights-tile.component.css
