+++ app/availability/app-insights/dependencies/app-insights-dependencies.component.html
