+++ app/availability/app-insights/exceptions/app-insights-exceptions.component.html
